requirements.txt
python-telegram-bot==22.6
requests>=2.31.0

deep-translator>=1.11.0
langdetect>=1.0.9

openai-whisper>=20231117
SpeechRecognition>=3.10.0
pydub>=0.25.1

python-docx>=1.1.0
PyPDF2>=3.0.0

pytesseract>=0.3.10
Pillow>=10.0.0

gTTS>=2.5.1
